refactor: 💡 infoText component -> text component
