+++ src/components/inputField/inputField.styles.scss
@@ -19,6 +19,10 @@
 				}
 			}
 		}
+		.formWrapper__infoText {
+			margin-top: 12px;
+			margin-bottom: 12px;
+		}
 	}
 
 	&__input {
