+++ src/components/profile/profile.styles.scss
@@ -470,7 +470,7 @@ $maxFormElementWidth: 320px;
 		}
 
 		.askerRegistration__consultingModeInfo {
-			padding: 0 $grid-base-three 0 20px;
+			padding: 0 $grid-base-three 40px 20px;
 		}
 
 		.selectedAgencyInfo,
